+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-01-31
+
+### Added
+- View menu toggles for status bar and line numbers
+- "+" tab button for quick new tabs
+- Save All option that saves non-empty tabs with txt default
+
+### Fixed
+- Edit > Paste uses clipboard text directly
+
 ## [0.2.6] - 2026-01-31
 
 ### Fixed